+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - *(cli)* Handle multiple instances of filecrabs
 - *(cli)* Add a command to `add` instances to a filecrab config
 - *(cli)* Add a command to `remove` an exisiting instance from a filecrab config
+- *(cli)* Add init command, to initialize filecrab's config without running anything
 
 ### 🐛 Bug Fixes